Exception Handling

+++ OrderSystem.Models/Order.cs
@@ -3,7 +3,7 @@
     public class Order
     {
         public long OrderId { get; set; }
-        public long MagicNumber { get; set; }
+        public int MagicNumber { get; set; }
         public string OrderText { get; set; }
     }
 }
